% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checksum.R
\name{checksum}
\alias{checksum}
\title{Prints a 'checksum' of datasets stored in text files}
\usage{
checksum()
}
\arguments{
\item{none}{}
}
\value{
Prints a description
}
\description{
Prints a 'checksum' of datasets stored in text files which are stored in a
  pipe-delimited format. It requires a shell environment variable called "CHECKSUM_DIR" which
  should contain the name of the directory where the text files are stored.
}
\examples{
checksum()
}

