% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadPaths.R
\name{loadPaths}
\alias{loadPaths}
\title{Add "^PATHS" to R environment}
\usage{
loadPaths()
}
\description{
Creates string objects that hold contents of environment variables containing
"^PATHS" from Sys.getenv().
}
\examples{
\dontrun{
loadPaths()
}
}
\seealso{
\code{\link{loadPathsDirect}} to specify path locations.
}

