% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadPathsDirect.R
\name{loadPathsDirect}
\alias{loadPathsDirect}
\title{Add paths from files to R environment}
\usage{
loadPathsDirect(additionalPath = "",
  standardPaths = c("output/data/paths.txt", "lib/paths.txt"))
}
\arguments{
\item{additionalPath,}{standardPaths}
}
\value{
Creates string objects that hold contents of specified files, normally paths.txt
}
\description{
Add paths from files to R environment
}
\seealso{
loadPaths.R
}

