% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadPaths.R
\name{loadPaths}
\alias{loadPaths}
\title{Add "^PATHS" to R environment}
\usage{
loadPaths()
}
\arguments{
\item{none}{}
}
\description{
Creates string objects that hold contents of environment variables containing "^PATHS" from Sys.getenv().
}
\seealso{
loadPathsDirect.R
}

