% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadPathsDirect.R
\name{loadPathsDirect}
\alias{loadPathsDirect}
\title{Add paths from files to R environment}
\usage{
loadPathsDirect(additionalPaths = "",
  standardPaths = c("output/data/paths.txt", "lib/paths.txt"))
}
\arguments{
\item{additionalPath,}{standardPaths}
}
\value{
Creates string objects that hold contents of specified files, normally paths.txt
}
\description{
Add paths from files to R environment
}
\examples{
#Load paths in standardPaths
loadPathsDirect()

#Load paths in additionalPaths and standardPaths
loadPathsDirect(additionalPath = c("example/path/one.txt", "example/path/two.txt"))

#Load paths in additionalPaths
loadPathsDirect(additionalPaths = c("example/path/one.txt", "example/path/two.txt"),
                standardPaths = "")
}
\seealso{
\code{\link{loadPaths}} to load from "^PATHS"
}

